+++ test/test_credentials.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 import pytest
 
-import pyega3.pyega3 as pyega3
+from pyega3.credentials import Credentials
 import test.conftest as common
 
 
@@ -12,7 +12,7 @@ def test_load_username_and_password_from_credentials_file(mock_input_file):
     password = common.rand_str()
 
     with mock_input_file(json.dumps({"username": username, "password": password})) as credentials_file:
-        result = pyega3.Credentials.from_file(credentials_file)
+        result = Credentials.from_file(credentials_file)
         assert result.username == username
         assert result.password == password
         assert result.key is None
@@ -24,7 +24,7 @@ def test_when_credentials_file_has_no_password_ask_user_for_it(mock_input_file):
 
     with mock_input_file(json.dumps({"username": username})) as credentials_file, mock.patch("getpass.getpass",
                                                                                              return_value=password):
-        result = pyega3.Credentials.from_file(credentials_file)
+        result = Credentials.from_file(credentials_file)
         assert result.username == username
         assert result.password == password
         assert result.key is None
@@ -33,7 +33,7 @@ def test_when_credentials_file_has_no_password_ask_user_for_it(mock_input_file):
 def test_error_when_credentials_file_is_bad_json(mock_input_file):
     with mock_input_file("bad json") as credentials_file:
         with pytest.raises(SystemExit):
-            pyega3.Credentials.from_file(credentials_file)
+            Credentials.from_file(credentials_file)
 
 
 def test_get_credential_prompts_user_for_username_and_password():
@@ -42,7 +42,7 @@ def test_get_credential_prompts_user_for_username_and_password():
 
     with mock.patch('builtins.input', return_value=username):
         with mock.patch('getpass.getpass', return_value=password):
-            credentials = pyega3.Credentials()
+            credentials = Credentials()
             credentials.prompt_for_missing_values()
             assert credentials.username == username
             assert credentials.password == password
@@ -56,7 +56,7 @@ def test_load_credential_prompts_user_for_credentials_if_credentials_file_does_n
     with mock.patch('builtins.input', return_value=username):
         with mock.patch('getpass.getpass', return_value=password):
             with mock.patch('os.path.exists', return_value=False):
-                credentials = pyega3.Credentials.from_file("unknownfile.txt")
+                credentials = Credentials.from_file("unknownfile.txt")
                 assert credentials.username == username
                 assert credentials.password == password
                 assert credentials.key is None
