+++ test/test_delete_temporary_files.py
@@ -4,13 +4,11 @@ import re
 import pytest
 import responses
 
+from pyega3.credentials import Credentials
 from pyega3 import pyega3 as pyega3
 
 test_file_id = 'test_file_id1'
 expected_file_size = pyega3.DOWNLOAD_FILE_SLICE_CHUNK_SIZE * 3
-
-
-
 
 
 def test_deleting_non_existent_file_does_not_raise_exception():
@@ -38,8 +36,9 @@ def test_temp_files_are_deleted_automatically_if_there_are_no_exceptions(mock_se
     input_file = bytearray(os.urandom(file_size_without_iv))
     mock_requests.add(responses.GET, f'{mock_server_config.url_api}/files/{test_file_id}', body=input_file, status=200)
 
-    pyega3.download_file_retry(pyega3.Credentials(), test_file_id, temporary_output_file, temporary_output_file,
-                               file_size_with_iv, 'check_sum', 1, temporary_output_file, None, 2, 0.1, mock_server_config)
+    pyega3.download_file_retry(Credentials(), test_file_id, temporary_output_file, temporary_output_file,
+                               file_size_with_iv, 'check_sum', 1, temporary_output_file, None, 2, 0.1,
+                               mock_server_config)
 
     temp_file = pyega3.TEMPORARY_FILES.pop()
     # The temporary file should not exist because everything went fine,
@@ -70,7 +69,7 @@ def download_with_exception(mock_requests, output_file_path, mock_server_config)
         mock_requests.add(responses.GET, f'{mock_server_config.url_api}/files/{test_file_id}', body=content, status=200)
 
     with pytest.raises(Exception) as context_manager:
-        pyega3.download_file_retry(pyega3.Credentials(), test_file_id, output_file_path, output_file_path,
+        pyega3.download_file_retry(Credentials(), test_file_id, output_file_path, output_file_path,
                                    expected_file_size, 'check_sum', 1, output_file_path, None,
                                    number_of_retries, 0.1, mock_server_config)
