+++ test/test_openid.py
@@ -4,6 +4,7 @@ from urllib import parse
 import pytest
 import responses
 
+from pyega3.credentials import Credentials
 import pyega3.pyega3 as pyega3
 import test.conftest as common
 
@@ -37,12 +38,12 @@ def mock_openid_server(mock_requests, mock_server_config):
 
 
 def test_get_token_from_openid_server(mock_openid_server, mock_server_config):
-    good_credentials = pyega3.Credentials(username=mock_openid_server.username, password=mock_openid_server.password)
+    good_credentials = Credentials(username=mock_openid_server.username, password=mock_openid_server.password)
     resp_token = pyega3.get_token(good_credentials, mock_server_config)
     assert resp_token == mock_openid_server.access_token
 
 
 def test_bad_openid_credentials_exits(mock_openid_server, mock_server_config):
-    bad_credentials = pyega3.Credentials(username=common.rand_str(), password=common.rand_str())
+    bad_credentials = Credentials(username=common.rand_str(), password=common.rand_str())
     with pytest.raises(SystemExit):
         pyega3.get_token(bad_credentials, mock_server_config)
