+++ test/test_retry_download.py
@@ -2,6 +2,7 @@ import os
 
 import responses
 
+from pyega3.credentials import Credentials
 from pyega3 import pyega3 as pyega3
 
 
@@ -28,7 +29,7 @@ def test_failed_download_retries_from_where_it_stopped(temporary_output_file, mo
                       status=200)
     mock_requests.add(responses.GET, f'{mock_server_config.url_api}/files/{file_id}', body=rest_of_the_input_file, status=200)
 
-    pyega3.download_file_retry(pyega3.Credentials(), file_id, temporary_output_file, temporary_output_file,
+    pyega3.download_file_retry(Credentials(), file_id, temporary_output_file, temporary_output_file,
                                file_size_with_iv, 'check_sum', 1, temporary_output_file, None, 2, 0.1, mock_server_config)
 
     assert mock_requests.calls[1].request.headers.get('Range') == 'bytes=0-92699'
